+++ labgrid/driver/quartushpsdriver.py
@@ -10,6 +10,7 @@ from ..resource.udev import AlteraUSBBlaster
 from ..step import step
 from .common import Driver, check_file
 from .exception import ExecutionError
+from ..util.managedfile import ManagedFile
 
 
 @target_factory.reg_driver
@@ -63,8 +64,8 @@ class QuartusHPSDriver(Driver):
     def flash(self, filename=None, address=0x0):
         if filename is None and self.image is not None:
             filename = self.target.env.config.get_image_path(self.image)
-        filename = os.path.abspath(os.path.expanduser(filename))
-        check_file(filename, command_prefix=self.interface.command_prefix)
+        mf = ManagedFile(filename)
+        mf.sync_to_resource()
 
         assert isinstance(address, int)
 
@@ -73,6 +74,6 @@ class QuartusHPSDriver(Driver):
         cmd += [
             "--cable={}".format(cable_number),
             "--addr=0x{:X}".format(address),
-            "--operation=P {}".format(filename),
+            "--operation=P {}".format(mf.get_remote_path()),
         ]
         subprocess.check_call(cmd)