+++ labgrid/driver/serialdriver.py
@@ -11,6 +11,7 @@ from ..protocol import ConsoleProtocol
 from ..resource import SerialPort, NetworkSerialPort
 from .common import Driver
 from .consoleexpectmixin import ConsoleExpectMixin
+from ..util.proxy import proxymanager
 
 
 @target_factory.reg_driver
@@ -53,10 +54,11 @@ class SerialDriver(ConsoleExpectMixin, Driver, ConsoleProtocol):
             self.serial.port = self.port.port
             self.serial.baudrate = self.port.speed
         else:
+            host, port = proxymanager.get_host_and_port(self.port, proxy=True)
             if self.port.protocol == "rfc2217":
-                self.serial.port = "rfc2217://{}:{}/".format(self.port.host, self.port.port)
+                self.serial.port = "rfc2217://{}:{}/".format(host, port)
             elif self.port.protocol == "raw":
-                self.serial.port = "socket://{}:{}/".format(self.port.host, self.port.port)
+                self.serial.port = "socket://{}:{}/".format(host, port)
             else:
                 raise Exception("SerialDriver: unknown protocol")
             self.serial.baudrate = self.port.speed