+++ labgrid/util/helper.py
@@ -1,0 +1,16 @@
+import os
+from socket import gethostname, socket, AF_INET, SOCK_STREAM
+from contextlib import closing
+
+def get_free_port():
+    """Helper function to always return an unused port."""
+    with closing(socket(AF_INET, SOCK_STREAM)) as s:
+        s.bind(('', 0))
+        return s.getsockname()[1]
+
+def get_user():
+    user = os.environ.get("USER")
+    if user:
+        return user
+    import pwd
+    return pwd.getpwuid(os.getuid())[0]